Fix the extension API tests.
