+++ src/test/api.functional.test.ts
@@ -3,16 +3,12 @@
 
 'use strict';
 
-// tslint:disable:no-any
+// tslint:disable:no-any max-func-body-length
 
 import { expect } from 'chai';
-import * as fs from 'fs';
-import * as glob from 'glob';
 import * as path from 'path';
 import { anyString, anything, instance, mock, when } from 'ts-mockito';
 import { buildApi } from '../client/api';
-import { ApplicationEnvironment } from '../client/common/application/applicationEnvironment';
-import { IApplicationEnvironment } from '../client/common/application/types';
 import { EXTENSION_ROOT_DIR } from '../client/common/constants';
 import { ExperimentsManager } from '../client/common/experiments';
 import { IExperimentsManager } from '../client/common/types';
@@ -22,8 +18,7 @@ import { ServiceContainer } from '../client/ioc/container';
 import { ServiceManager } from '../client/ioc/serviceManager';
 import { IServiceContainer, IServiceManager } from '../client/ioc/types';
 
-// tslint:disable-next-line: max-func-body-length
-suite('Extension API Debugger', () => {
+suite('Extension API - Debugger', () => {
     const expectedLauncherPath = `${EXTENSION_ROOT_DIR.fileToCommandArgument()}/pythonFiles/ptvsd_launcher.py`;
     const ptvsdPath = path.join('path', 'to', 'ptvsd');
     const ptvsdHost = 'somehost';
@@ -40,12 +35,12 @@ suite('Extension API Debugger', () => {
         experimentsManager = mock(ExperimentsManager);
         debugAdapterFactory = mock(DebugAdapterDescriptorFactory);
 
-        when(serviceManager.get(IExperimentsManager))
+        when(serviceContainer.get<IExperimentsManager>(IExperimentsManager))
             // Return the mock.
-            .thenReturn(Promise.resolve(instance(experimentsManager)));
-        when(serviceManager.get(IDebugAdapterDescriptorFactory))
+            .thenReturn(instance(experimentsManager));
+        when(serviceContainer.get<IDebugAdapterDescriptorFactory>(IDebugAdapterDescriptorFactory))
             // Return the mock.
-            .thenReturn(Promise.resolve(instance(debugAdapterFactory)));
+            .thenReturn(instance(debugAdapterFactory));
     });
 
     function mockInExperiment(host: string, port: number, wait: boolean) {
@@ -155,63 +150,5 @@ suite('Extension API Debugger', () => {
         const expectedArgs = [ptvsdPath, '--host', ptvsdHost, '--port', ptvsdPort.toString(), '--wait'];
 
         expect(args).to.be.deep.equal(expectedArgs);
-    });
-});
-
-suite('Extension version tests', () => {
-    let version: string;
-    let applicationEnvironment: IApplicationEnvironment;
-    const branchName = process.env.CI_BRANCH_NAME;
-
-    suiteSetup(async function() {
-        // Skip the entire suite if running locally
-        if (!branchName) {
-            // tslint:disable-next-line: no-invalid-this
-            return this.skip();
-        }
-    });
-
-    setup(() => {
-        applicationEnvironment = new ApplicationEnvironment(undefined as any, undefined as any, undefined as any);
-        version = applicationEnvironment.packageJson.version;
-    });
-
-    test('If we are running a pipeline in the master branch, the extension version in `package.json` should have the "-dev" suffix', async function() {
-        if (branchName !== 'master') {
-            // tslint:disable-next-line: no-invalid-this
-            return this.skip();
-        }
-
-        return expect(
-            version.endsWith('-dev'),
-            'When running a pipeline in the master branch, the extension version in package.json should have the -dev suffix'
-        ).to.be.true;
-    });
-
-    test('If we are running a pipeline in the release branch, the extension version in `package.json` should not have the "-dev" suffix', async function() {
-        if (!branchName!.startsWith('release')) {
-            // tslint:disable-next-line: no-invalid-this
-            return this.skip();
-        }
-
-        return expect(
-            version.endsWith('-dev'),
-            'When running a pipeline in the release branch, the extension version in package.json should not have the -dev suffix'
-        ).to.be.false;
-    });
-});
-
-suite('Extension localization files', () => {
-    test('Load localization file', () => {
-        const filesFailed: string[] = [];
-        glob.sync('package.nls.*.json', { sync: true, cwd: EXTENSION_ROOT_DIR }).forEach(localizationFile => {
-            try {
-                JSON.parse(fs.readFileSync(path.join(EXTENSION_ROOT_DIR, localizationFile)).toString());
-            } catch {
-                filesFailed.push(localizationFile);
-            }
-        });
-
-        expect(filesFailed).to.be.lengthOf(0, `Failed to load JSON for ${filesFailed.join(', ')}`);
     });
 });
